+++ net/include/net/os_tunnel.h
@@ -15,6 +15,7 @@
 #include "common/error.h"
 #include "common/socket_address.h"
 #include "common/utils.h"
+#include "net/network_manager.h"
 #include "vpn/platform.h"
 #include "vpn/utils.h"
 
@@ -209,7 +210,7 @@ protected:
     uint32_t m_if_index = 0;
     bool m_system_dns_setup_success = false;
     bool m_ipv6_available = false;
-    bool m_tunnel_active_owned = false;
+    VpnTunnelActivityToken m_tunnel_activity_token = VPN_TUNNEL_ACTIVITY_TOKEN_INVALID;
 };
 
 #ifdef __linux__